requirements.txt
fastapi==0.119.0
uvicorn==0.37.0
python-dotenv==1.1.1
psycopg[binary]==3.2.10
pydantic==2.12.2
starlette==0.48.0
anyio==4.11.0
typing_extensions==4.15.0
